Habits rotes

+++ habit_tracker_service/src/services/habits_service.py
@@ -1,12 +1,12 @@
-'''Сервис для работы с пользователями'''
+'''Сервис для работы с привычками'''
 
 
 from fastapi import HTTPException, status
 from typing import List
 
-from habit_tracker_service.src.services.base_habits_service import AbstractHabitsService
+from src.services.base_habits_service import AbstractHabitsService
 from src.repositories.base_repository import AbstractRepository
-from src.schemas.habits_schemas import HabitsCredsSchema
+from src.schemas.habits_schemas import HabitCredsSchema, HabitSchema
 from src.models.habits_model import HabitsModel
 # from src.logging.habits_service_# logger import UsersService# logger
 
@@ -14,12 +14,13 @@ from src.models.habits_model import HabitsModel
 # logger = UsersService# logger().get_# logger()
 
 
-class UsersService(AbstractHabitsService):
+class HabitsService(AbstractHabitsService):
     '''Сервис для работы с реальными привычками'''
     def __init__(self, habits_repository: AbstractRepository):
         self.habits_repository: AbstractRepository = habits_repository()
-    
-    async def create(self, data: HabitsCredsSchema) -> int:
+
+
+    async def create(self, data: HabitCredsSchema) -> int:
         '''Создаёт привычку и возвращает её id'''
         # logger.info('Creating new habit name: %s', data.name)
 
@@ -35,10 +36,10 @@ class UsersService(AbstractHabitsService):
         return habit_id
     
 
-    async def get_by_id(self, habit_id: int) -> HabitsModel:
+    async def get_by_id(self, user_id: int, habit_id: int) -> HabitsModel:
         '''Возвращает данные привычки по id'''
         # logger.info('Retrieving the habit, id: %d', habit_id)
-        habits = await self.habits_repository.get_by({'id': habit_id})
+        habits = await self.habits_repository.get_by({'id': habit_id, 'user_id': user_id})
 
         if len(habits) == 0:
             # logger.warning('User id: %d not found', habit_id)
@@ -48,15 +49,25 @@ class UsersService(AbstractHabitsService):
         return habits[0]
 
 
-    async def get_all(self) -> List[HabitsModel]:
+    async def get_all(self, user_id: int) -> List[HabitsModel]:
         '''Возвращает все привычки'''
-        habits = await self.habits_repository.get_by({})
+        habits = await self.habits_repository.get_by({'user_id': user_id})
         return habits
 
 
-    async def update(self, habit_id: int, new_data: HabitsCredsSchema):
+    async def update(self, user_id: int, habit_id: int, new_data: HabitCredsSchema):
         '''Обновляет данные привычки'''
         # logger.info('Updating the habit, id: %d', habit_id)
+
+        habits = await self.habits_repository.get_by({'id': habit_id})
+
+        if len(habits) == 0:
+            # logger.warning('User id: %d not found', habit_id)
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Habit not found')
+
+        if habits[0].user_id != user_id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Habit of another user')
+
         habit_dict = new_data.model_dump()
         try:
             await self.habits_repository.update(habit_id, habit_dict)
@@ -66,9 +77,19 @@ class UsersService(AbstractHabitsService):
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
 
 
-    async def delete(self, habit_id: int):
+    async def delete(self, user_id: int, habit_id: int):
         '''Удаляет привычку'''
         # logger.info('Deleting the habit, id: %d', habit_id)
+
+        habits = await self.habits_repository.get_by({'id': habit_id})
+
+        if len(habits) == 0:
+            # logger.warning('User id: %d not found', habit_id)
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Habit not found')
+
+        if habits[0].user_id != user_id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Habit of another user')
+
         try:
             await self.habits_repository.delete(habit_id)
             # logger.info('User id: %d successfully deleted', habit_id)
